Don't use ipaddress backport for parsing IP addresses.
There are multiple versions of the ipaddress backport from python3 to
python2 which all seem to have some missing features or completely
different behaviour at some places.
(best example: `ipaddress.ip_network('192.168.148.0/28')` results in
an IPv6 network with one version of the module)

All those modules share the name "ipaddress" and it is really hard to
determine which one is currently in use.

Because of this problem I replaced the lines which used the ipaddress
module with a few regex and generator lines, which should be okay as the
data format in those places is defined by the MIB and shouldn't change.